% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myImagePlot}
\alias{myImagePlot}
\title{Matrix plotting}
\usage{
myImagePlot(x, main, ...)
}
\arguments{
\item{x}{a \code{matrix} to be plotted.}

\item{main}{header text for the plot.}

\item{...}{additional arguments (not implemented).}
}
\description{
Plot a heatmap with colorbar.
}
\examples{
myImagePlot(matrix(1:12,3,4), 'A header')

}
\references{
T. Mehmood, K.H. Liland, L. Snipen, S. Sæbø, A review of variable selection 
methods in Partial Least Squares Regression, Chemometrics and Intelligent Laboratory Systems
118 (2012) 62-69.
}
\seealso{
\code{\link{VIP}} (SR/sMC/LW/RC), \code{\link{filterPLSR}}, \code{\link{shaving}}, 
\code{\link{stpls}}, \code{\link{truncation}},
\code{\link{bve_pls}}, \code{\link{ga_pls}}, \code{\link{ipw_pls}}, \code{\link{mcuve_pls}},
\code{\link{rep_pls}}, \code{\link{spa_pls}},
\code{\link{lda_from_pls}}, \code{\link{lda_from_pls_cv}}, \code{\link{setDA}}.
}
\author{
Tahir Mehmood, Kristian Hovde Liland, Solve Sæbø.
}
