% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setDA}
\alias{setDA}
\title{Set chosen Discriminant Analysis}
\usage{
setDA(LQ = NULL)
}
\arguments{
\item{LQ}{character argument 'lda', 'qda', 'max' or NULL}
}
\value{
Returns the default set method.
}
\description{
The default methods is LDA, but QDA and column of maximum prediction can be chosen.
}
\examples{
\dontrun{
setDA() # Query 'lda', 'qda' or 'max'
setDA('qda') # Set default method to QDA
}
}
\seealso{
\code{\link{VIP}} (SR/sMC/LW/RC), \code{\link{filterPLSR}}, \code{\link{shaving}}, 
\code{\link{stpls}}, \code{\link{truncation}},
\code{\link{bve_pls}}, \code{\link{ga_pls}}, \code{\link{ipw_pls}}, \code{\link{mcuve_pls}},
\code{\link{rep_pls}}, \code{\link{spa_pls}},
\code{\link{lda_from_pls}}, \code{\link{lda_from_pls_cv}}, \code{\link{setDA}}.
}

