% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrV.R
\name{summary.mvrV}
\alias{summary.mvrV}
\title{Summary method for stpls and trunc}
\usage{
\method{summary}{mvrV}(object, what = c("all", "validation", "training"),
  digits = 4, print.gap = 2, ...)
}
\arguments{
\item{object}{an mvrV object}

\item{what}{one of "all", "validation" or "training"}

\item{digits}{integer. Minimum number of significant digits in the output. Default is 4.}

\item{print.gap}{Integer. Gap between coloumns of the printed tables.}

\item{...}{Other arguments sent to underlying methods.}
}
\description{
Adaptation of \code{summary.mvr} from the \code{pls} package v 2.4.3.
}
\seealso{
\code{\link[pls]{summary.mvr}}
}

