% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.R
\name{truncation}
\alias{truncation}
\title{Trunction PLS}
\usage{
truncation(..., Y.add, weights, method = "truncation")
}
\arguments{
\item{...}{arguments passed on to \code{mvrV}).}

\item{Y.add}{optional additional response vector/matrix found in the input data.}

\item{weights}{optional object weighting vector.}

\item{method}{choice (default = \code{truncation}).}
}
\value{
Returns an object of class mvrV, simliar to to mvr object of the pls package.
}
\description{
Distribution based truncation for variable selection in subspace 
methods for multivariate regression.
}
\details{
Loading weights are truncated around their median based on confidence intervals
for modelling without replicates (Lenth et al.).
}
\examples{
data(yarn, package = "pls")
tr <- truncation(density ~ NIR, ncomp=5, data=yarn, validation="CV",
 truncation="Lenth", trunc.width=0.95) # Default truncation
summary(tr)

}
\author{
Kristian Hovde Liland.
}
\references{
K.H. Liland, M. H?y, H. Martens, S. S?b?: Distribution based truncation for 
variable selection in subspace methods for multivariate regression, Chemometrics and
Intelligent Laboratory Systems 122 (2013) 103-111.
}
\seealso{
\code{\link{VIP}} (SR/sMC/LW/RC), \code{\link{filterPLSR}}, \code{\link{spa_pls}}, 
\code{\link{stpls}}, \code{\link{trunc}}, \code{\link{bve_pls}}, \code{\link{mcuve_pls}},
\code{\link{ipw_pls}}, \code{\link{ga_pls}}, \code{\link{rep_pls}}.
}

